Upd: some json encoding and decoding enhancements

+++ src/lib/jsonrpc.c
@@ -386,7 +386,7 @@ bool json_iterate_object(sds s, const char *path, iterate_callback icb, void *ic
             return false;
         }
         if (klen > 2) {
-            if (sds_json_unescape(p + koff + 1, klen - 2, &key) == false ||
+            if (sds_json_unescape(p + koff + 1, (size_t)(klen - 2), &key) == false ||
                 vcb_isalnum(value) == false)
             {
                 _set_parse_error(error, "Validation of key in path \"%s\" has failed. Key must be alphanumeric.", path);
@@ -404,7 +404,7 @@ bool json_iterate_object(sds s, const char *path, iterate_callback icb, void *ic
         switch(vtype) {
             case MJSON_TOK_STRING:
                 if (vlen > 2) {
-                    if (sds_json_unescape(p + voff + 1, vlen - 2, &value) == false) {
+                    if (sds_json_unescape(p + voff + 1, (size_t)(vlen - 2), &value) == false) {
                         _set_parse_error(error, "JSON unescape error for value for key \"%s\" in JSON path \"%s\" has failed", key, path);
                         FREE_SDS(value);
                         FREE_SDS(key);
@@ -564,7 +564,7 @@ static bool _json_get_string(sds s, const char *path, size_t min, size_t max, sd
     n = n - 2;
     p++;
 
-    if ((sds_json_unescape(p, n, result) == false) ||
+    if ((sds_json_unescape(p, (size_t)n, result) == false) ||
         (sdslen(*result) < min || sdslen(*result) > max))
     {
         _set_parse_error(error, "Value length %lu for JSON path \"%s\" is out of bounds", sdslen(*result), path);
